feat(children): bones with nested layout: children :rocket:
Refactoring in order to extract a new type that describe a bone with children.
