+++ src/index.tsx
@@ -1,4 +1,4 @@
-import { Bone } from './Bone';
+import { Bone } from './Bone/Bone';
 import { Skeleton } from './Skeleton';
 
 export { Bone, Skeleton };
